# swagger/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "DanceManager API",
    "description": "댄스매니저의 데이터를 서빙합니다.",
    "version": "1.0.0"
  },
  "host": "localhost:80",
  "basePath": "/",
  "schemes": ["http"],
  "paths": {
    "/": {
      "get": {
        "tags": ["기본"],
        "description": "기본 인사 메시지",
        "responses": {
          "200": {
            "description": "성공",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "number",
                  "example": 200
                },
                "message": {
                  "type": "string",
                  "example": "안녕하세요 댄스매니저 API를 서빙해주고 있어요 ^ㅁ^"
                },
                "data": {
                  "type": "string",
                  "example": "https://swagger...."
                }
              }
            }
          }
        }
      }
    },
    "/music": {
      "get": {
        "tags": ["음악"],
        "description": "전체 음악 목록 조회",
        "responses": {
          "200": {
            "description": "성공",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "string"
                  },
                  "title": {
                    "type": "string"
                  },
                  "category": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "500": {
            "description": "서버 에러"
          }
        }
      }
    },
    "/music/{category}": {
      "get": {
        "tags": ["음악"],
        "description": "카테고리별 음악 목록 조회",
        "parameters": [
          {
            "name": "category",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "음악 카테고리"
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "string"
                  },
                  "title": {
                    "type": "string"
                  },
                  "category": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "500": {
            "description": "서버 에러"
          }
        }
      }
    }
  }
}